requirements.txt
africastalking==1.2.9
asgiref==3.9.1
Authlib==1.6.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
dotenv==0.9.9
drf-jwt==1.19.2
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.1
phonenumberslite==9.0.11
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.5
